% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.ivppmlhdfe}
\alias{summary.ivppmlhdfe}
\title{Summarize an IV-PPML-HDFE fit}
\usage{
\method{summary}{ivppmlhdfe}(object, eform = FALSE, irr = FALSE, ...)
}
\arguments{
\item{object}{an `"ivppmlhdfe"` fit.}

\item{eform, irr}{logical; display exponentiated coefficients
(incidence-rate ratios). `irr` is an alias for `eform`.}

\item{...}{unused.}
}
\description{
Summarize an IV-PPML-HDFE fit
}
